JECC - Diagramme - Summen stimmen nicht #389 JECC - Diagramme - der Zeitbereich unter dem Diagramm zeigt immer (?) 15 Minuten zuviel an #393 JECC - Diagramme - es scheint nicht möglich zu sein verschiende Datentypen in ein Diagramm zu laden #382 Neue Einheiten + Umrechnung in Diagrammen benötigt [Prio Hoch] #367 JECC - Diagramme - Aggregation im Lade-Dialog #387 JECC - Diagramme - fixed Error with zeros in data for maximum calculation JECC - Charts - default TimeFrames #385 improved JECalc for big data rows fixed spelling error
