+++ JEApplication/src/main/java/org/jevis/application/Chart/Charts/LineChart.java
@@ -39,7 +39,7 @@ public class LineChart implements Chart {
     private static SaveResourceBundle rb = new SaveResourceBundle(AppLocale.BUNDLE_ID, AppLocale.getInstance().getLocale());
     private static final Logger logger = LogManager.getLogger(AreaChart.class);
     private String chartName;
-    private String unit;
+    private List<String> unit = new ArrayList<>();
     private List<ChartDataModel> chartDataModels;
     private Boolean hideShowIcons;
     private ObservableList<XYChart.Series<Number, Number>> series = FXCollections.observableArrayList();
@@ -66,6 +66,8 @@ public class LineChart implements Chart {
                 hexColors.add(singleRow.getColor());
                 series.add(serie.getSerie());
                 tableData.add(serie.getTableEntry());
+                String currentUnit = UnitManager.getInstance().formate(singleRow.getUnit());
+                if (!unit.contains(currentUnit)) unit.add(currentUnit);
 
             } catch (JEVisException e) {
                 e.printStackTrace();
@@ -73,8 +75,7 @@ public class LineChart implements Chart {
         }
 
         if (chartDataModels != null && chartDataModels.size() > 0) {
-            unit = UnitManager.getInstance().formate(chartDataModels.get(0).getUnit());
-            if (unit.equals("")) unit = rb.getString("plugin.graph.chart.valueaxis.nounit");
+            if (unit.isEmpty()) unit.add(rb.getString("plugin.graph.chart.valueaxis.nounit"));
             period = chartDataModels.get(0).getAttribute().getDisplaySampleRate();
         }
 
@@ -92,13 +93,19 @@ public class LineChart implements Chart {
 
         lineChart.getXAxis().setAutoRanging(true);
 
-        Period period = new Period(getStartDateTime(), getEndDateTime().plus(getPeriod()).plusSeconds(1));
+        Period period = new Period(getStartDateTime(), getEndDateTime().plus(getPeriod()).plusSeconds(1)).minus(this.period);
         period = period.minusSeconds(period.getSeconds());
         String overall = period.toString(PeriodFormat.wordBased().withLocale(AppLocale.getInstance().getLocale()));
 
         lineChart.getXAxis().setLabel(rb.getString("plugin.graph.chart.dateaxis.title") + " " + overall);
+
         lineChart.getYAxis().setAutoRanging(true);
-        lineChart.getYAxis().setLabel(unit);
+        String allUnits = "";
+        for (String s : unit) {
+            if (unit.indexOf(s) == 0) allUnits += s;
+            else allUnits += ", " + s;
+        }
+        lineChart.getYAxis().setLabel(allUnits);
 
         initializeZoom();
     }
