+++ JEApplication/src/main/resources/jeapplication_de.properties
@@ -53,7 +53,7 @@ graph.selection.removeselections=Auswahl abw\u00e4hlen
 graph.table.color=Farbe
 graph.table.load=Auswahl
 graph.table.selectchart=Diagramm
-graph.table.interval=Interval
+graph.table.interval=Intervall
 graph.table.cleaning=Bereinigung
 graph.table.startdate=Start Datum
 graph.table.enddate=End Datum
