+++ JEVisControlCenter/src/main/resources/JEVisCC_de.properties
@@ -199,6 +199,14 @@ plugin.graph.dialog.export.decimalseparator.selection=Auswahl
 plugin.graph.progress.message=Importiere Messwerte . . .
 plugin.graph.progress.header=Importiert Messwerte..
 plugin.graph.progress.title=Importiere Messwerte
+plugin.graph.interval.label=Intervall
+plugin.graph.interval.preset=Vorgabe
+plugin.graph.interval.hourly=St\u00fcndlich
+plugin.graph.interval.daily=T\u00e4glich
+plugin.graph.interval.weekly=W\u00f6chentlich
+plugin.graph.interval.monthly=Monatlich
+plugin.graph.interval.quarterly=Viertelj\u00e4hrlich
+plugin.graph.interval.yearly=J\u00e4hrlich
 plugin.object.attribute.target.button=Ziel
 plugin.object.attribute.target.goto=Gehe zu
 plugin.object.attribute.target.goto.tooltip=Gehe zu Ziel im Baum
