+++ JEVisControlCenter/src/main/resources/JEVisCC_en.properties
@@ -203,6 +203,14 @@ plugin.graph.dialog.export.decimalseparator.selection=Selection
 plugin.graph.progress.message=Importing Values . . .
 plugin.graph.progress.header=Importing Values..
 plugin.graph.progress.title=Importing Values
+plugin.graph.interval.label=Interval
+plugin.graph.interval.preset=Preset
+plugin.graph.interval.hourly=Hourly
+plugin.graph.interval.daily=Daily
+plugin.graph.interval.weekly=Weekly
+plugin.graph.interval.monthly=Monthly
+plugin.graph.interval.quarterly=Quarterly
+plugin.graph.interval.yearly=Yearly
 plugin.object.attribute.target.button=Target
 plugin.object.attribute.target.goto=Goto
 plugin.object.attribute.target.goto.tooltip=Goto target in tree
